Allow SPARK_APP_ID to be present as a global constant for all SparkUI.
SparkUI - Executors page makes 3 ajax calls to present the page
1. Get the `ApplicationId`
2. Fetch the `executorspage-template.html`
3. Get `allexecutors` to populate the executors and summary datatable.

This PR will make the first two ajax calls unnecessary.
A. Add the `ApplicationId` at the top of the page as a global constant like the below, making the first ajax call unnecassary.
```
<script>const SPARK_APP_ID='application_1502369394594_0032';</script>
```
B. Convert the `executorspage-template.html` to a JS file, and include it as a regular javascript file, so it is browser-cached.