[SPARK-30410][SQL][2.4] Calculating size of table with large number of partitions causes flooding logs
### What changes were proposed in this pull request?

Backported from [pr#27079](https://github.com/apache/spark/pull/27079). For a partitioned table, if the number of partitions are very large, e.g. tens of thousands or even larger, calculating its total size causes flooding logs.
The flooding happens in:
1. `calculateLocationSize` prints the starting and ending for calculating the location size, and it is called per partition;
2. `bulkListLeafFiles` prints all partition paths.

This pr is to simplify the logging when calculating the size of a partitioned table.

### How was this patch tested?

not related

Closes #27143 from wzhfy/improve_log-2.4.

Authored-by: Zhenhua Wang <wzh_zju@163.com>
Signed-off-by: Takeshi Yamamuro <yamamuro@apache.org>