[SPARK-31446][WEBUI][3.0] Make html elements for a paged table possible to have different id attribute
### What changes were proposed in this pull request?

This PR backports #28217 to `branch-3.0`, making each id attribute for page navigations in a page unique.

PagedTable#pageNavigation returns HTML elements representing a page navigation for a paged table.
In the current implementation, the method generates an id and it's used for id attribute for a set of elements for the page navigation.
But some pages have two page navigations so there are two set of elements where corresponding elements have the same id.
For example, there are two form-completedJob-table-page id in JobsPage.

### Why are the changes needed?

Each id attribute should be unique in a page.

### Does this PR introduce any user-facing change?

No.

### How was this patch tested?

I added a test case for pageNavigation extended.
I also manually tested that there were no warning messages for the uniqueness in JobsPage and JobPage.

Closes #28246 from sarutak/backporting-SPARK-31446-branch-3.0.

Authored-by: Kousuke Saruta <sarutak@oss.nttdata.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>