[SPARK-31234][SQL][2.4] ResetCommand should not affect static SQL Configuration
This PR is to backport the fix of https://github.com/apache/spark/pull/28003, add a migration guide, update the PR description and add an end-to-end test case.

Before this PR, the SQL `RESET` command will reset the values of static SQL configuration to the default and remove the cached values of Spark Context Configurations in the current session. This PR fixes the bugs. After this PR, the `RESET` command follows its definition and only updates the runtime SQL configuration values to the default.

When we introduced the feature of Static SQL Configuration, we did not update the implementation of  SQL `RESET` command.

The static SQL configuration should not be changed by any command at runtime. However, the `RESET` command resets the values to the default. We should fix them.

Before Spark 2.4.6, the `RESET` command resets both the runtime and static SQL configuration values to the default. It also removes the cached values of Spark Context Configurations in the current session, although these configuration values are for displaying/querying only.

Added an end-to-end test and a unit test

Closes #28262 from gatorsmile/spark-31234followup2.4.

Authored-by: gatorsmile <gatorsmile@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>