rdar://69319047 Spark should reject invalid CatalogImplementation Configuration
Fail fast when SessionStateBuilder is set but not used.

There are several ways configuration can be incorrect and we can cover more of them with
fast failure rather than hitting strange errors later.

Yes, previously if SessionStateBuilder was set but not used, no error would be thrown. Now it is invalid
to have a configuration with this parameter set but not used.

New unit test added