[SPARK-30894][SQL][2.4] Make Size's nullable independent from SQL config changes
This is a backport of https://github.com/apache/spark/pull/27658

### What changes were proposed in this pull request?

In the PR, I propose to add the `legacySizeOfNull ` parameter to the `Size` expression, and pass the value of `spark.sql.legacy.sizeOfNull` if `legacySizeOfNull` is not provided on creation of `Size`.

### Why are the changes needed?

This allows to avoid the issue when the configuration change between different phases of planning, and this can silently break a query plan which can lead to crashes or data corruption.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

By `CollectionExpressionsSuite`.

Closes #30058 from anuragmantri/SPARK-30894-2.4.

Authored-by: Maxim Gekk <max.gekk@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>