[SPARK-26646][TEST][PYSPARK][2.4] Fix flaky test: pyspark.mllib.tests.StreamingLogisticRegressionWithSGDTests.test_training_and_prediction
### What changes were proposed in this pull request?

This is backport of SPARK-26646 to branch-2.4 to fix flaky test in the branch.

### Why are the changes needed?

The test pyspark.mllib.tests.StreamingLogisticRegressionWithSGDTests.test_training_and_prediction looks sometimes flaky.

```
Traceback (most recent call last):
  File "/home/runner/work/spark/spark/python/pyspark/mllib/tests.py", line 1492, in test_training_and_prediction
    self._eventually(condition, timeout=180.0)
  File "/home/runner/work/spark/spark/python/pyspark/mllib/tests.py", line 133, in _eventually
    lastValue = condition()
  File "/home/runner/work/spark/spark/python/pyspark/mllib/tests.py", line 1487, in condition
    self.assertGreater(errors[1] - errors[-1], 0.3)
AssertionError: -0.07000000000000006 not greater than 0.3
```

The predict stream can possibly be consumed to the end before the input stream. When it happens, the model improvement is not high as expected and causes test failed. This patch tries to increase number of batches of streams. This won't increase test time because we have a timeout there.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Unit test

Closes #30078 from viirya/SPARK-26646-2.4.

Authored-by: Liang-Chi Hsieh <viirya@gmail.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>