Add stored procedure API (#791)
### What changes were proposed in this pull request?

This PR adds a stored procedure API. For now, this PR is submitted against 3.0.1-apple directly. I'll write a design doc and submit this change for discussion early next year.

### Why are the changes needed?

This change is needed for our Iceberg integration in Spark 3.

### Does this PR introduce _any_ user-facing change?

Yes but this change is isolated and won't affect anyone.

### How was this patch tested?

This PR comes with a few tests. More tests in Iceberg.