[MINOR] Lazily instantiate Hive classes
### What changes were proposed in this pull request?

Hive classes should be lazily instantiated in `HiveDelegationTokenProvider`

### Why are the changes needed?

The instantiation of reflection classes within the constructor results in a connection attempt to secure metastore which fails when --proxy-user is used since this initialization happens **_before_** the delegation token for proxy user is obtained. This results in exactly one failed attempt at connecting to metastore before eventually succeeding.

Using `lazy val` ensures that these classes are not instantiated until methods of this class are called. After this change, metastore connectivity succeeds in one attempt. 

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Tested on dev cluster with secure hive metastore