[SPARK-34271][SQL] Use majorMinorPatchVersion for Hive version parsing
Use `majorMinorPatchVersion` to check major & minor version in `IsolatedClientLoader.hiveVersion`.

Currently `IsolatedClientLoader.hiveVersion` needs to enumerate all Hive patch versions. Therefore, whenever we upgrade Hive version we'd need to remember to update the method as well. It would be better if we just check major & minor version.

No.

This is a refactoring and relies on existing tests.

Closes #31371 from sunchao/replace-hive-version.

Authored-by: Chao Sun <sunchao@apple.com>
Signed-off-by: HyukjinKwon <gurwls223@apache.org>