[SPARK-34316][K8S] Support spark.kubernetes.executor.disableConfigMap
This PR aims to add a new configuration `spark.kubernetes.executor.disableConfigMap`.

This can be use to disable config map creating for executor pods due to https://github.com/apache/spark/pull/27735 .

No. By default, this doesn't change AS-IS behavior.
This is a new feature to add an ability to disable SPARK-30985.

Pass the newly added UT.

Closes #31428 from dongjoon-hyun/SPARK-34316.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>