[SPARK-34359][SQL][3.1] Add a legacy config to restore the output schema of SHOW DATABASES
This backports https://github.com/apache/spark/pull/31474 to 3.1/3.0

This is a followup of https://github.com/apache/spark/pull/26006

In #26006 , we merged the v1 and v2 SHOW DATABASES/NAMESPACES commands, but we missed a behavior change that the output schema of SHOW DATABASES becomes different.

This PR adds a legacy config to restore the old schema, with a migration guide item to mention this behavior change.

Improve backward compatibility

No (the legacy config is false by default)

a new test

Closes #31486 from cloud-fan/command-schema.

Authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>