[SPARK-34201][SQL][TESTS][3.0] Fix the test code build error for docker-integration-tests
### What changes were proposed in this pull request?

This PR fixes the test code build error for `docker-integration-tests` in `branch-3.0`.
Currently, building test code for docker-integration-tests in `branch-3.0` seems to fail due to `ojdbc6:11.2.0.1.0` and `jdb2jcc4:10.5.0.5` are not available.

The solution backports SPARK-32049 and partially SPARK-31272.

### Why are the changes needed?

To fix the build error.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

This change affects `OracleIntegrationSuite`  so I confirmed it passed.
This change also affects `DB2IntegrationSuite` but it's ignored for now so I didn't care for it.

Closes #31291 from sarutak/fix-test-build-error.

Authored-by: Kousuke Saruta <sarutak@oss.nttdata.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>