Add Iceberg back (#985)
* Add Iceberg as a dep

* rdar://70004937 Rewrite row-level operations for Iceberg (#922)

* rdar://72811621 Control distribution and ordering during table creation (#939)

This PR adds support for defining distribution and ordering during table creation.

This change is needed for feature parity with ADT in Spark 2.

This PR adds new optional table creation clauses but it should only affect Iceberg users.

This PR comes with new tests. More tests are in Iceberg.

Updates Iceberg to 0.12.0-preview-7 (#1018)