[SPARK-34855][CORE] Avoid local lazy variable in SparkContext.getCallSite
### What changes were proposed in this pull request?

`SparkContext.getCallSite` uses local lazy variable. In Scala 2.11, local lazy val requires synchronization so for large number of job submissions in the same context, it will be a bottleneck. This only for branch-2.4 as we drop Scala 2.11 support at SPARK-26132.

### Why are the changes needed?

To avoid possible bottleneck for large number of job submissions in the same context.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing tests.

Closes #31988 from viirya/SPARK-34855.

Authored-by: Liang-Chi Hsieh <viirya@gmail.com>
Signed-off-by: HyukjinKwon <gurwls223@apache.org>