[SPARK-36129][BUILD][3.1] Upgrade commons-compress to 1.21 to deal with CVEs
This PR backports the change of SPARK-36129 (#33333) which upgrades `commons-compress` from `1.20` to `1.21` to deal with CVEs.

Some CVEs which affect `commons-compress 1.20` are reported and fixed in `1.21`.
https://commons.apache.org/proper/commons-compress/security-reports.html

* CVE-2021-35515
* CVE-2021-35516
* CVE-2021-35517
* CVE-2021-36090

The severities are reported as low for all the CVEs but it would be better to deal with them just in case.

No.

CI.

Closes #33338 from sarutak/backport-SPARK-36129-branch-3.1.

Authored-by: Kousuke Saruta <sarutak@oss.nttdata.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>